requirements.txt
click==8.1.7
python-dotenv==1.0.0
pymongo==4.6.1
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.12.0
langchain-core==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.8
slack-bolt==1.18.0